==> Script3_Joining.ipynb <==
# Joining 

The purpose of this notebook is to join the cleaned traffic and collisions tables in such a way that for each traffic entry, we have a column which states how many collisions have occurred nearby.  It takes in the two files `traffic_cleaned_by_hour.csv` and `collisions_cleaned.csv`, which can be obtained by running the `Traffic_Cleaning` and `Collisions_Cleaning` notebooks.  The final joined table is saved here as `joined_by_hour.csv`.

We begin by importing the standard libraries, and reading in the datasets.  We also define a function `distance`, which calculates the geographical distance between two points, given two pairs of latitude-longitude coordinates.

import pandas as pd
import numpy as np

from math import cos, asin, sqrt, pi

# function which calculates distance between two lat-lon pairs, distance returned in kms
def distance(lat1, lon1, lat2, lon2):
    p = pi/180
    a = 0.5 - ((lat2-lat1)*p).apply(cos)/2 + (lat1*p).apply(cos) * (lat2*p).apply(cos) * (1-((lon2-lon1)*p).apply(cos))/2
    return 12742 * ((a).apply(sqrt)).apply(asin) #2*R*asin...

traffic = pd.read_csv('Data/traffic_cleaned_by_hour.csv')
collisions = pd.read_csv('Data/collisions_cleaned.csv')

traffic.head()

collisions.head()

# sort dataset by intersection and date-time
df = traffic
df = df.sort_values(['Id_Intersection', 'Year', 'Month', 'Day', 'hour'])

# merge with the collisions table (left join)
print('Merging')
merged = df.merge(collisions, how='left', left_on=['Year', 'Month', 'Day'], right_on=['year', 'month', 'day'])

# calculate distance between the traffic count and the collision
print('Calculating distances')
merged['dist'] = distance(merged['Latitude'], 
                      merged['Longitude'],
                      merged['lat'], 
                      merged['lon'])

# add a column to state whether the collision is nearby
merged['collision_nearby'] = np.where((merged['dist'] < 2) & \
                                   (np.abs(merged['hour_x'] - merged['hour_y']) < 3), 1, 0)

# take only the columns we need
print('Filtering')
merged = merged[['Id_Intersection', 'Year', 'Month', 'Day', 'hour_x', 'collision_nearby']]

# aggregate on the number of collisions nearby (sum)
print('Grouping and Aggregating')
totals = merged.groupby(['Id_Intersection', 
                       'Year', 'Month', 'Day', 'hour_x']).sum()['collision_nearby'].reset_index()

# recombine with the original subset table
print('Recombining')
new_traffic = df.merge(totals, how='left', left_on=['Id_Intersection', 'Year', 'Month', 'Day', 'hour'], 
                        right_on=['Id_Intersection', 'Year', 'Month', 'Day', 'hour_x'])

new_traffic.drop('hour_x', axis=1, inplace=True)

# save the recombined table
print('Saving')
new_traffic.to_csv('Data/joined_by_hour.csv', index=False)
