==> Script5_Predicting.ipynb <==
import numpy as np
import pandas as pd

joined = pd.read_csv('Data/Montreal/Final/joined_by_hour.csv')
new = pd.read_csv('Data/Montreal/Final/joined_by_hour.csv')
collisions = pd.read_csv('Data/Montreal/Final/collisions_cleaned.csv')

joined.head()

collisions.head()

# binarize collision_nearby
new['is_collision'] = np.where(new['collision_nearby'] > 0, 1, 0)

# drop some non-useful columns
new.drop(columns=['Id_Intersection', 'Date', 'collision_nearby'], inplace=True)

# check results
new.head()

# import the different sklearn packages we will need
from sklearn.model_selection import train_test_split
from sklearn.preprocessing import StandardScaler
from sklearn.preprocessing import MinMaxScaler

from sklearn.linear_model import LogisticRegression
from sklearn.tree import DecisionTreeClassifier
from sklearn.svm import LinearSVC

from sklearn.pipeline import Pipeline
from sklearn.model_selection import GridSearchCV
import joblib
import warnings
warnings.filterwarnings('ignore')

# ensure that the dataset is sorted by date-time
new.sort_values(['Year', 'Month', 'Day', 'hour'], inplace=True)
new.head()

# split into features and target
X = new.iloc[:, :-1]
y = new['is_collision']

# train test split with the shuffle set to False
X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=99, shuffle=False)

dt = DecisionTreeClassifier(max_depth=5)
dt.fit(X_train, y_train);

dt.score(X_train, y_train)

dt.score(X_test, y_test)

So at this point, we have the trained model.  What I need is
1. The lat/lon values of the intersections which we use
2. The lat/lon values of the collisions which occurred nearby
3. The probabilities for each intersection which we use

- `test_collisions` gives me the locations of the collisions
- 

test_rows = new.loc[(new['Year'] == 2018) & (new['Month'] == 2) & (new['Day'] == 14) \
                      & (new['hour'] == 8), :]
test_rows

X_test_rows = test_rows.iloc[:, :-1]
y_test_rows = test_rows['is_collision']

probs = dt.predict_proba(X_test_rows)
ones = [round(x[1], 2) for x in probs]

test_rows['prob'] = ones
test_rows['pred'] = np.where(test_rows['prob'] > 0.5, 1, 0)
test_rows

test_rows.to_csv('../Data/Montreal/Final/predictions_intersections.csv', index=False)

test_collisions = collisions.loc[(collisions['year'] == 2018) & \
                                 (collisions['month'] == 2) & \
                                 (collisions['day'] == 14) & \
                                 (collisions['hour'] > 5) & (collisions['hour'] < 11), :]

test_collisions.head()

test_collisions.shape

test_collisions.to_csv('../Data/Montreal/Final/predictions_collisions.csv', index=False)

pd.set_option('max_columns', None)

test_collisions['Longitude'] = test_collisions['lon']
test_collisions['Latitude'] = test_collisions['lat']
test_collisions.head()

test_collisions['is_intersection'] = 0
test_rows['is_intersection'] = 1

test_collisions['colour_scheme'] = 1
test_rows['colour_scheme'] = np.where(test_rows['pred'] == 1, 2, 3)

test = pd.concat([test_rows, test_collisions])

test

test.to_csv('../Data/Montreal/Final/predictions.csv', index=False)

test['L']
